Feat: Added util numWithComma func
천단위 콤마 출력 함수를 util에 옮김
